// src/pages/Home/types.ts
export interface ContainerProps {
  border?: boolean
  children: React.ReactNode
}

export interface ButtonProps {
  color?: string
  fixedWidth?: boolean
  name?: string
  children: React.ReactNode
  onClick?: () => void
}

export interface SvgIconProps {
  src: string
  width: string
  height: string
}

export interface InputProps {
  name: string
  placeholder: string
  type?: string
  value?: string
  onChange: (event: React.ChangeEvent<HTMLInputElement> | React.ChangeEvent<HTMLTextAreaElement>) => void
}

export interface validateProps {
  name: string
  message: string
  email: string
}

export interface ContactProps {
  title: string
  content: string
  id: string
}

export interface ValidationTypeProps {
  type: any
}